+++ src/components/App/spec.tsx
@@ -15,16 +15,20 @@ describe('App', () => {
 
     it('should add/remove the .active class from Card components on mousedown', () => {
       const wrapper = mount(<App />);
+
       wrapper.find('.card').at(0).simulate('mousedown');
       expect(wrapper.find('.card').at(0).hasClass('active')).toBe(true);
+
       wrapper.find('.card').at(0).simulate('mousedown');
       expect(wrapper.find('.card').at(0).hasClass('active')).toBe(false);
     });
 
     it('should remove the .active class from Card components when App is clicked', () => {
       const wrapper = mount(<App />);
+
       wrapper.find('.card').at(0).simulate('mousedown');
       expect(wrapper.find('.card').at(0).hasClass('active')).toBe(true);
+
       const selector = wrapper.find('.app').at(0).prop('onClick');
       const event = {} as React.MouseEvent;
       if (selector) act(() => selector(event));
@@ -38,10 +42,12 @@ describe('App', () => {
     "End Turn" ArrowIcon is subsequently clicked`, () => {
       const wrapper = mount(<App />);
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+
       const selector = wrapper.find('.stackedCard').at(0).prop('onClick');
       if (selector) act(() => selector(stopPropagationMouseEvent));
       wrapper.update();
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
       const endTurnSelector = wrapper.find('.endTurnIcon').at(0).prop('onClick');
       if (endTurnSelector) act(() => endTurnSelector(stopPropagationMouseEvent));
       wrapper.update();
@@ -51,8 +57,9 @@ describe('App', () => {
     it(`should perform no action if the "End Turn" ArrowIcon is clicked while the "Take a Card"
     ArrowIcon is already showing`, () => {
       const wrapper = mount(<App />);
-      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
       const event = {} as React.MouseEvent;
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+
       const endTurnSelector = wrapper.find('.endTurnIcon').at(0).prop('onClick');
       if (endTurnSelector) act(() => endTurnSelector(event));
       wrapper.update();
@@ -61,8 +68,9 @@ describe('App', () => {
 
     it('should perform no action if an ArrowIcon with class drawCardIcon is clicked', () => {
       const wrapper = mount(<App />);
-      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
       const event = {} as React.MouseEvent;
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+
       const drawCardSelector = wrapper.find('.drawCardIcon').at(0).prop('onClick');
       if (drawCardSelector) act(() => drawCardSelector(event));
       wrapper.update();
@@ -73,13 +81,80 @@ describe('App', () => {
     is not showing`, () => {
       const wrapper = mount(<App />);
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+
       const deckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
       if (deckSelector) act(() => deckSelector(stopPropagationMouseEvent));
       wrapper.update();
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
       if (deckSelector) act(() => deckSelector(stopPropagationMouseEvent));
       wrapper.update();
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+    });
+  });
+
+  describe('handCards state', () => {
+    it('should not draw a Card if handCards has 5 items', () => {
+      const wrapper = mount(<App />);
+      const event = {} as React.MouseEvent;
+
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+
+      // draw first card
+      const deckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (deckSelector) act(() => deckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
+      // draw second card
+      const endTurnSelector = wrapper.find('.endTurnIcon').at(0).prop('onClick');
+      if (endTurnSelector) act(() => endTurnSelector(event));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+      const secondDeckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (secondDeckSelector) act(() => secondDeckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
+      // // draw third card
+      const secondEndTurnSelector = wrapper.find('.endTurnIcon').at(0).prop('onClick');
+      if (secondEndTurnSelector) act(() => secondEndTurnSelector(event));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+      const thirdDeckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (thirdDeckSelector) act(() => thirdDeckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
+      // // draw fourth card
+      const thirdEndTurnSelector = wrapper.find('.endTurnIcon').at(0).prop('onClick');
+      if (thirdEndTurnSelector) act(() => thirdEndTurnSelector(event));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+      const fourthDeckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (fourthDeckSelector) act(() => fourthDeckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
+      // // draw fifth card
+      const fourthEndTurnSelector = wrapper.find('.endTurnIcon').at(0).prop('onClick');
+      if (fourthEndTurnSelector) act(() => fourthEndTurnSelector(event));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+      const fifthDeckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (fifthDeckSelector) act(() => fifthDeckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
+      // // attempt to draw a sixth card
+      const fifthEndTurnSelector = wrapper.find('.endTurnIcon').at(0).prop('onClick');
+      if (fifthEndTurnSelector) act(() => fifthEndTurnSelector(event));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+      const sixthDeckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (sixthDeckSelector) act(() => sixthDeckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
     });
   });
 
@@ -87,8 +162,10 @@ describe('App', () => {
     it('should not hide the Hand component if it\'s already displaying and it\'s clicked', () => {
       const wrapper = mount(<App />);
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
+
       wrapper.find('.handIcon').at(0).simulate('click');
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
+
       wrapper.find('.hand').at(0).simulate('click');
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
     });
@@ -97,8 +174,10 @@ describe('App', () => {
     clicked`, () => {
       const wrapper = mount(<App />);
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
+
       wrapper.find('.handIcon').at(0).simulate('click');
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
+
       wrapper.find('.card').at(0).simulate('mousedown');
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
     });
@@ -106,9 +185,11 @@ describe('App', () => {
     it('should not display the drawCardIcon on the Deck once the deck is clicked', () => {
       const wrapper = mount(<App />);
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+
       const selector = wrapper.find('.stackedCard').at(0).prop('onClick');
       if (selector) act(() => selector(stopPropagationMouseEvent));
       wrapper.update();
+
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
     });
 
@@ -116,6 +197,7 @@ describe('App', () => {
       const wrapper = mount(<App />);
       expect(wrapper.find('.hand').find('.hidden').at(0)).toHaveLength(1);
       expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(1);
+
       const selector = wrapper.find('.stackedCard').at(0).prop('onClick');
       if (selector) act(() => selector(stopPropagationMouseEvent));
       wrapper.update();
@@ -126,10 +208,12 @@ describe('App', () => {
     showing`, () => {
       const wrapper = mount(<App />);
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
+
       const deckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
       if (deckSelector) act(() => deckSelector(stopPropagationMouseEvent));
       wrapper.update();
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
+
       const appSelector = wrapper.find('div').at(0).prop('onClick');
       if (appSelector) act(() => appSelector(stopPropagationMouseEvent));
       wrapper.update();
@@ -139,10 +223,12 @@ describe('App', () => {
     it('should show the hand when the HandIcon is clicked', () => {
       const wrapper = mount(<App />);
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
+
       const selector = wrapper.find('.handIcon').find('div').at(0).prop('onClick');
       if (selector) act(() => selector(stopPropagationMouseEvent));
       wrapper.update();
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
+
       const updatedSelector = wrapper.find('.handIcon').find('div').at(0).prop('onClick');
       if (updatedSelector) act(() => updatedSelector(stopPropagationMouseEvent));
       wrapper.update();
@@ -163,8 +249,10 @@ describe('App', () => {
     clicked`, () => {
       const wrapper = mount(<App />);
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
+
       wrapper.find('.handIcon').at(0).simulate('click');
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
+
       wrapper.find(App).simulate('click');
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
     });